refactor(engine): remove dead code and de-duplicate audit findings
Removes unused API surface (Move::is_en_passant, MoveList::as_mut_slice,
IntoIterator for &MoveList, Bitboard::{FULL,RANK_1,RANK_8},
CastleRights::remove), merges the two duplicate splitmix64 PRNGs into a
shared rng module, implements Display/FromStr for Square instead of
shadowing inherent methods, fixes two clippy nits, and de-duplicates the
now_ms test helper into tests/common.

+++ engine/src/position.rs
@@ -196,7 +196,7 @@ impl Position {
             MoveFlag::Quiet => self.move_piece(from, to),
             MoveFlag::DoublePush => {
                 self.move_piece(from, to);
-                let ep_sq = Square(((from.0 + to.0) / 2) as u8);
+                let ep_sq = Square((from.0 + to.0) / 2);
                 self.ep = Some(ep_sq);
                 self.hash ^= zobrist::ep_file(ep_sq.file());
             }
